+++ realm/realm-library/src/main/java/io/realm/exceptions/RealmMigrationNeededException.java
@@ -21,7 +21,7 @@ import java.io.File;
 import io.realm.internal.Keep;
 
 @Keep
-public class RealmMigrationNeededException extends RuntimeException {
+public final class RealmMigrationNeededException extends RuntimeException {
 
     private final String canonicalRealmPath;